requirements.txt
numpy>=1.20
tqdm>=4.60
rich>=11.2.0
pyrealsense2>=2.50.0.3812
setuptools>=60.9.3
opencv-python>=4.5.5.62
uvicorn>=0.20.0
fastapi>=0.87.0
pyyaml>=6.0
flask>=2.0.2
gevent>=21.12.0